+++ Assets/CleanArchitecture_Example/Scripts/Domain/UseCases/IPurchaseBundleUseCase.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace CleanArchitecture_Example.Scripts.Domain.UseCases
+{
+    public interface IPurchaseBundleUseCase
+    {
+        void Purchase(int bundleId, Action<int> onSucceed, Action onFailed);
+    }
+}
