+++ Assets/CleanArchitecture_Example/Scripts/Domain/UseCases/IShowLoadingScreenUseCase.cs
@@ -1,4 +1,4 @@
-namespace CleanArchitecture_Example.Scripts.Domain
+namespace CleanArchitecture_Example.Scripts.Domain.UseCases
 {
     public interface IShowLoadingScreenUseCase
     {
