+++ Assets/CleanArchitecture_Example/Scripts/Domain/UseCases/IShowShopUseCaseOutput.cs.meta
