+++ Assets/CleanArchitecture_Example/Scripts/Domain/UseCases/PurchaseBundleUseCase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b0a2a08bf1c70f14ea097375c84de3dc
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
