+++ Assets/CleanArchitecture_Example/Scripts/Domain/UseCases/ShowShopUseCase.cs
@@ -1,4 +1,4 @@
-namespace CleanArchitecture_Example.Scripts.Domain
+namespace CleanArchitecture_Example.Scripts.Domain.UseCases
 {
     public class ShowShopUseCase : IShowShopUseCase
     {
