+++ Assets/CleanArchitecture_Example/Scripts/InterfaceAdapters/LoadingScreenPresenter.cs
@@ -1,4 +1,5 @@
 using CleanArchitecture_Example.Scripts.Domain;
+using CleanArchitecture_Example.Scripts.Domain.UseCases;
 
 namespace CleanArchitecture_Example.Scripts.InterfaceAdapters
 {
