+++ Assets/CleanArchitecture_Example/Scripts/InterfaceAdapters/Services/EndpointHelper.cs.meta
