+++ Assets/CleanArchitecture_Example/Scripts/InterfaceAdapters/Services/IImageRepository.cs
@@ -1,6 +1,6 @@
 using UnityEngine;
 
-namespace CleanArchitecture_Example.Scripts.InterfaceAdapters
+namespace CleanArchitecture_Example.Scripts.InterfaceAdapters.Services
 {
     public interface IImageRepository
     {
